+++ source/Core/Routing/Module/ClassProviderStorage.php
@@ -11,6 +11,7 @@ use OxidEsales\Eshop\Core\Registry;
 /**
  * Handler class for the storing of the metadata controller field of the modules.
  *
+ * @deprecated since v6.4.0 (2019-03-22); Use service 'OxidEsales\EshopCommunity\Internal\Common\Storage\YamlFileStorage'.
  * @internal Do not make a module extension for this class.
  * @see      http://wiki.oxidforge.org/Tutorials/Core_OXID_eShop_classes:_must_not_be_extended
  */